requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langchain-chroma
pypdf
python-docx
reportlab
pandas
openpyxl
striprtf
Pillow
pytesseract
pillow-heif
ddgs
PyYAML
